[IOTMID-87] Adds initial docs structure
This adds an initial proposed structure for some of the documents
to be written as to compose the internal beta release of the
platform.

This initial proposal includes: an architecture document, an operations
guide and an user guide. For each of those, only the overall introduction
is presented.

Apart from those, it is also proposeed that the new components
implemented in order to support the platform and set it apart from
the "stock" fiware implementation and its usage on IoT scenarios.
Those however, would have their structure defined by the components
themselves, and would also be generated and versioned on either
the components own version control system, or a new segregate git
repository.

Change-Id: I4be39ec522b6c8f44ab20106cc1705038b25a186